+++ client/master.jsx
@@ -22,7 +22,7 @@ export default class Master extends Component {
                   {id:"phonetic",state:true,name:'Phonetic'},
                   {id:"ngram",state:false,name:'Partial'},
                   {id:"normalized",state:true,name:'Normalized'},
-                  {id:"translation",state:false,name:'Translation'}
+                  {id:"translation",state:false,name:'Translation (beta)'}
                 ]},
         {id:"Surah",state:true,name:'Chapter Names',options: [
                   {id:"phonetic",state:true,name:'Phonetic'},
@@ -62,7 +62,7 @@ export default class Master extends Component {
         {id:"EnglishCorpus",state:false,name:'English - corpus',options: [
                   {id:"ngram",state:true,name:'Partial'},
                   {id:"normalized",state:true,name:'Normalized'},
-                  {id:"translation",state:true,name:'Translation'}
+                  {id:"translation",state:true,name:'Translation (beta)'}
                 ]}
 
       ],